moved request and cookie parser into separate files, perform session refresh if cookies are expired
