+++ integrations/authentication/parseCookies.js
@@ -1,0 +1,8 @@
+export default function parseCookies(response) {
+  const raw = response.headers.raw()['set-cookie'];
+  return raw.map((entry) => {
+    const parts = entry.split(';');
+    const cookiePart = parts[0];
+    return cookiePart;
+  }).join(';');
+}
